// frontend/src/api/models/listSalesOrdersStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListSalesOrdersStatus =
  (typeof ListSalesOrdersStatus)[keyof typeof ListSalesOrdersStatus]

export const ListSalesOrdersStatus = {
  draft: 'draft',
  confirmed: 'confirmed',
  shipped: 'shipped',
  completed: 'completed',
  cancelled: 'cancelled',
} as const